Refactor textFeedback and add a time's up message
