+++ app/src/main/java/ch/epfl/sweng/SDP/game/drawing/DrawingOnline.java
@@ -25,6 +25,7 @@ import ch.epfl.sweng.SDP.localDatabase.LocalDbHandlerForImages;
 import ch.epfl.sweng.SDP.matchmaking.GameStates;
 import ch.epfl.sweng.SDP.matchmaking.Matchmaker;
 
+import static ch.epfl.sweng.SDP.game.drawing.FeedbackTextView.timeIsUpTextFeedback;
 import static java.lang.String.format;
 
 public class DrawingOnline extends GyroDrawingActivity {
@@ -63,6 +64,7 @@ public class DrawingOnline extends GyroDrawingActivity {
                 GameStates stateEnum = GameStates.convertValueIntoState(state);
                 switch (stateEnum) {
                     case WAITING_UPLOAD:
+                        timeIsUpTextFeedback(getApplicationContext());
                         uploadDrawing().addOnCompleteListener(
                                 new OnCompleteListener<TaskSnapshot>() {
                                     @Override
